# docs/spikes/S7-api-v1/recorded/swagger_v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MintPlayer API",
    "description": "API that allows access to the data of MintPlayer. ",
    "version": "v1"
  },
  "paths": {
    "/api/v1/Account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Registers a new MintPlayer user.",
        "operationId": "api-account-register",
        "requestBody": {
          "description": "Registration data for the new user.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/UserDataVM"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/UserDataVM"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/UserDataVM"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDataVM"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDataVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserDataVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserDataVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/Account/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Attempts to login a user and generates a Bearer token for that user.",
        "operationId": "api-account-login",
        "requestBody": {
          "description": "Login credentials for the user.",
          "content": {
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginVM"
              }
            },
            "text/xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginVM"
              }
            },
            "application/*+xml": {
              "schema": {
                "$ref": "#/components/schemas/LoginVM"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginVM"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginVM"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginVM"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginVM"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Subject/{subject_id}/likes": {
      "get": {
        "tags": [
          "Subject"
        ],
        "summary": "Get the number of likes/dislikes for a subject.",
        "operationId": "api-subject-getlikes",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "description": "Id of the person/artist/song.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Subject"
        ],
        "summary": "Like/dislike a specific subject.",
        "operationId": "api-subject-like",
        "parameters": [
          {
            "name": "subject_id",
            "in": "path",
            "description": "Id of the person/artist/song.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Like = true, dislike = false.",
          "content": {
            "application/xml": {
              "schema": {
                "type": "boolean"
              }
            },
            "text/xml": {
              "schema": {
                "type": "boolean"
              }
            },
            "application/*+xml": {
              "schema": {
                "type": "boolean"
              }
            },
            "application/json-patch+json": {
              "schema": {
                "type": "boolean"
              }
            },
            "application/json": {
              "schema": {
                "type": "boolean"
              }
            },
            "text/json": {
              "schema": {
                "type": "boolean"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "boolean"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              },
              "text/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectLikeResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LoginResult": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LoginStatus"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "error": {
            "type": "string",
            "nullable": true
          },
          "errorDescription": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoginStatus": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "LoginVM": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubjectLikeResult": {
        "type": "object",
        "properties": {
          "likes": {
            "type": "integer",
            "format": "int32"
          },
          "dislikes": {
            "type": "integer",
            "format": "int32"
          },
          "like": {
            "type": "boolean",
            "nullable": true
          },
          "authenticated": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "isTwoFactorEnabled": {
            "type": "boolean"
          },
          "bypass2faForExternalLogin": {
            "type": "boolean"
          },
          "pictureUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserDataVM": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "passwordConfirmation": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}